+++ modules/core/dependency/python-ihm/make-release.sh
@@ -21,4 +21,5 @@ mv src/ihm_format_wrap.c "src/ihm_format_wrap_${VERSION}.c"
 python3 setup.py sdist
 rm -f "src/ihm_format_wrap_${VERSION}.c"
 
-echo "Now use 'twine upload dist/ihm-${VERSION}.tar.gz' to publish the release on PyPi"
+echo "Now use 'twine upload dist/ihm-${VERSION}.tar.gz' to publish the release on PyPi."
+echo "Then, update the conda-forge and Homebrew packages to match."